Modify the nRF5x timer and system init code to use the HFXO manager.
This allows the timer code to interoperate with other code that uses the HFXO (such as BLE) such that the HFXO is only enabled when required and turned off when not in use, saving power. The nrf5X_clock_hfxo_wait_until_started function has also been added, which allows code to wait until the HFXO has started up before continuing.